+++ astroquery/heasarc/tests/test_heasarc.py
@@ -274,12 +274,26 @@ def test_locate_data():
         Heasarc.locate_data()
 
     with pytest.raises(
-        TypeError, match="query_result need to be an astropy.table.Table"
+        TypeError, match=(
+            "query_result need to be an astropy.table.Table or astropy.table.Row"
+        )
     ):
         Heasarc.locate_data([1, 2])
 
     with pytest.raises(ValueError, match="No __row column found"):
         Heasarc.locate_data(Table({"id": [1, 2, 3.0]}), catalog_name="xray")
+
+
+def test_locate_data_row():
+    table = Table({"id": [1, 2, 3.0]})
+
+    # we cannot do full call as this is not remote,
+    # but if we check that we error on __row not input type
+    with pytest.raises(ValueError, match="No __row column found"):
+        Heasarc.locate_data(table[0], catalog_name="xray")
+
+    with pytest.raises(ValueError, match="No __row column found"):
+        Heasarc.locate_data(table[0:2], catalog_name="xray")
 
 
 def test_download_data__empty():
@@ -330,6 +344,25 @@ def test_download_data__outside_sciserver():
         Heasarc.download_data(
             Table({"sciserver": ["some-link"]}), host="sciserver"
         )
+
+
+def test_download_data__table_row():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        datadir = f'{tmpdir}/data'
+        downloaddir = f'{tmpdir}/download'
+        os.makedirs(datadir, exist_ok=True)
+        with open(f'{datadir}/file.txt', 'w') as fp:
+            fp.write('data')
+        # include both a file and a directory
+        tab = Table({'sciserver': [f'{tmpdir}/data/file.txt', f'{tmpdir}/data']})
+        # The patch is to avoid the test that we are on sciserver
+        with patch('os.path.exists') as exists:
+            exists.return_value = True
+            Heasarc.download_data(tab[0], host="sciserver", location=downloaddir)
+            Heasarc.download_data(tab[1], host="sciserver", location=downloaddir)
+        assert os.path.exists(f'{downloaddir}/file.txt')
+        assert os.path.exists(f'{downloaddir}/data')
+        assert os.path.exists(f'{downloaddir}/data/file.txt')
 
 
 # S3 mock tests